Move noise_model out of pulser.backends
